First approach to Alarms section
Added all the required UI and the logic for trigger alarms. As explained
in the comments of alarms.py I'm using the linux signal ALRM, I don't
really know if this is a good idea or not but we will figure out, hehe.

The handlers of the alarms are pending to do.