Refactored some of the code. Using Class + staticmethod for creating descriptive namespaces
Programming is not just making something work, but make the code readeble
self-descriptive and easy to maintain. In the helpers functions I have a lot
of non related funcionts and that's why I'm wrapping the functions into
a Class and using staticmethod. I think the code is more readable this way.