I couldn't help myself. Check permissions and owner of the session files.
If an evil user that have access to the RPi but not root acces, knows Python,
and creates a fake file of a session under /tmp, and sets the sessions
cookie exactly the same as the fake file UUID, he/she could get access to a
RaspCTL session without event have to make login. In he/she would deserve it!

Seriously, the probability of this happening is the same as being adopted
by Angelina Jolie the same day you do two pentakills playing at LoL while
touching yourself.

As I said, I couldn't help myself. Love and [code] ugliness!